+++ src/components/Header/Header.jsx
@@ -4,34 +4,26 @@ import { IoSearchOutline } from "react-icons/io5";
 import { FaStar } from "react-icons/fa";
 import { Logo, BackLink } from "../../components";
 import { pageTitle } from "../../helpers";
+import { useWidth } from "../../hooks";
 import { TextField, IconRouteLink, PageTitle } from "../../components";
+import { SEARCH_PAGE_SHORT_TITLE } from "../../pages/constans";
 import styles from "./Header.module.scss";
 import variables from "../../styles/variables.scss";
-
-const TITLE_SEARCH_SHORT = "Search Users";
 
 const Header = ({ onGetQuery, location, isSearchPage, isUserPage }) => {
   const title = pageTitle(location, isUserPage);
   const [query, setQuery] = useState("");
-  const [width, setWidth] = useState(0);
-  const showShortTitle = useMemo(
-    () => isSearchPage && width < variables.laptop,
-    [width]
-  );
-
-  const handleResize = useCallback(({ target }) => {
-    setWidth(target.innerWidth);
-  }, []);
-
-  useEffect(() => {
-    window.addEventListener("resize", handleResize);
-    return () => window.removeEventListener("resize", handleResize);
-  }, [window]);
+  const { width } = useWidth();
 
   const handleOnChange = useCallback(({ target: { value } }) => {
     setQuery(value);
     onGetQuery(value);
   }, []);
+
+  const showShortTitle = useMemo(
+    () => isSearchPage && width < variables.laptop,
+    [width]
+  );
 
   return (
     <header
@@ -47,7 +39,7 @@ const Header = ({ onGetQuery, location, isSearchPage, isUserPage }) => {
             [styles[`titleContainer--isSearch`]]: isSearchPage,
           })}
         >
-          <PageTitle title={showShortTitle ? TITLE_SEARCH_SHORT : title} />
+          <PageTitle title={showShortTitle ? SEARCH_PAGE_SHORT_TITLE : title} />
         </div>
       </div>
       {isSearchPage && (
