+++ src/hooks/index.js
@@ -3,3 +3,4 @@ export { default as useLocalStorage } from "./useLocalStorage";
 export { default as useFavorites } from "./useFavorites";
 export { default as useFetchCurrentUser } from "./useFetchCurrentUser";
 export { default as useScroll } from "./useScroll";
+export { default as useWidth } from "./useWidth";
