+++ src/styles/mixins.scss
@@ -32,7 +32,7 @@
 }
 
 @mixin breakpoint($width-property, $breakpoint) {
-  @media (#{$width-property}: $breakpoint) {
+  @media (#{$width-property}: #{$breakpoint}px) {
     @content;
   }
 }
